DISMTools 0.7.3 Preview 1 (#324)
* [PE Helper] Add script to change keyboard layout

* [Unattended answer file/DTSS] Add new File Extensions starter script

* [Refactor] More LINQ Usage (Part 1)

* [ServiceManager] Sort services in svchost groups by name before showing them

* [PE Helper] DTPE background override support

* [UserData] New UserData folder functionality

* [Unattended answer file/DTSS Browser] New Category for User scripts

* [Unattended answer file] Stop answer file creation with non-existent UnattendGen

* [Installer] Override ISL messages to introduce version fields

Closes #306

* [PE Helper/SysprepPreparator] Update binary

* [Unattended answer file/DTSS] Move /F parameter from the end to the beginning of REG ADD

* [REL] Update What's New section

+++ Helpers/extps1/PE_Helper/PE_Helper.ps1
@@ -606,19 +606,24 @@ function Start-PECustomization
                 icacls "$imagePath\Windows\system32\winpe.jpg" /grant "$(Get-LocalizedUsers -admins $true):(M)" | Out-Host
                 icacls "$imagePath\Windows\system32\winpe.jpg" /grant "$(Get-LocalizedUsers -admins $false):(M)" | Out-Host
                 Write-Host "Changing wallpaper..."
-                switch ($arch)
-                {
-                    x86 {
-                        Copy-Item -Path "$((Get-Location).Path)\backgrounds\winpe_x86.jpg" -Destination "$imagePath\Windows\system32\winpe.jpg" -Force
-                    }
-                    amd64 {
-                        Copy-Item -Path "$((Get-Location).Path)\backgrounds\winpe_amd64.jpg" -Destination "$imagePath\Windows\system32\winpe.jpg" -Force
-                    }
-                    arm64 {
-                        Copy-Item -Path "$((Get-Location).Path)\backgrounds\winpe_arm64.jpg" -Destination "$imagePath\Windows\system32\winpe.jpg" -Force
-                    }
-                    default {
-                        Copy-Item -Path "$((Get-Location).Path)\backgrounds\winpe_amd64.jpg" -Destination "$imagePath\Windows\system32\winpe.jpg" -Force
+                if (Test-Path -Path "$((Get-Location).Path)\backgrounds\wallpaper.jpg" -PathType Leaf) {
+                    # A wallpaper override has been detected. We'll use it instead.
+                    Copy-Item -Path "$((Get-Location).Path)\backgrounds\wallpaper.jpg" -Destination "$imagePath\Windows\system32\winpe.jpg" -Force
+                } else {
+                    switch ($arch)
+                    {
+                        x86 {
+                            Copy-Item -Path "$((Get-Location).Path)\backgrounds\winpe_x86.jpg" -Destination "$imagePath\Windows\system32\winpe.jpg" -Force
+                        }
+                        amd64 {
+                            Copy-Item -Path "$((Get-Location).Path)\backgrounds\winpe_amd64.jpg" -Destination "$imagePath\Windows\system32\winpe.jpg" -Force
+                        }
+                        arm64 {
+                            Copy-Item -Path "$((Get-Location).Path)\backgrounds\winpe_arm64.jpg" -Destination "$imagePath\Windows\system32\winpe.jpg" -Force
+                        }
+                        default {
+                            Copy-Item -Path "$((Get-Location).Path)\backgrounds\winpe_amd64.jpg" -Destination "$imagePath\Windows\system32\winpe.jpg" -Force
+                        }
                     }
                 }
                 Write-Host "Wallpaper changed"
@@ -729,6 +734,7 @@ function Start-PECustomization
                 Set-Content -Path "$imagePath\Windows\system32\startnet.cmd" -Value $contents -Force
             }
             Copy-Item -Path "$((Get-Location).Path)\files\startup\StartInstall.ps1" -Destination "$imagePath\StartInstall.ps1" -Force
+            Copy-Item -Path "$((Get-Location).Path)\files\startup\ChangeKeyboardLayout.ps1" -Destination "$imagePath\ChangeKeyboardLayout.ps1" -Force
             Copy-Item -Path "$((Get-Location).Path)\files\startup\DTPE_Inventory.ps1" -Destination "$imagePath\DTPE_Inventory.ps1" -Force
             Copy-Item -Path "$((Get-Location).Path)\files\dim_start\dimstart.bat" -Destination "$imagePath\dimstart.bat" -Force
             Copy-Item -Path "$((Get-Location).Path)\files\startup\menu.ps1" -Destination "$imagePath\menu.ps1" -Force
